feat: implémentation HttpOnly cookies pour sécurité JWT + refactoring services
Backend:
- HttpOnly cookies pour le token JWT (protection XSS)
- Modification AuthController: token dans cookie, endpoint /logout
- JwtAuthenticationFilter lit le token depuis cookie ou header
- Transformation services en interfaces + implémentations:
  * AuthService -> AuthServiceImpl
  * UserService -> UserServiceImpl
  * PostService -> PostServiceImpl
  * CommentService -> CommentServiceImpl
  * SubscriptionService -> SubscriptionServiceImpl
  * ThemeService -> ThemeServiceImpl
- Cookie avec attributs: HttpOnly, SameSite=Lax, Secure (prod)

Frontend:
- AuthService utilise withCredentials pour les cookies
- Suppression localStorage (plus sécurisé)
- Interceptor envoie automatiquement les cookies
- logout() appelle l'endpoint backend
- getCurrentUser() récupère infos depuis /me (Observable)
- Fix ThemesComponent pour getCurrentUserId() Observable
- Fix tests avec Observable

Sécurité:
- Token JWT inaccessible via JavaScript (protection XSS)
- Cookie HttpOnly + SameSite + Secure
- CORS configuré pour credentials
