+++ front/angular-app/src/app/features/themes/themes.component.spec.ts
@@ -208,7 +208,7 @@ describe('ThemesComponent', () => {
   });
 
   it('should not subscribe when user is not logged in', () => {
-    authService.getCurrentUserId.mockReturnValue(null);
+    authService.getCurrentUserId.mockReturnValue(of(null));
     const consoleSpy = jest.spyOn(console, 'error').mockImplementation();
     fixture.detectChanges();
